N-queens
Given N, and for a given N x N chessboard, found a way to place N queens such that no queen can attack any other queen on the chess board. A queen can be killed when it lies in the same row, or same column, or the same diagonal of any of the other queens.Printed all such configurations.